Add CI; rubocop cleanup; make generic for non-aws clients
