+++ .github/workflows/ci.yml
@@ -1,0 +1,32 @@
+name: Benchmark Tools CI
+
+on:
+  push:
+    branches:
+      - main
+
+  pull_request:
+    branches:
+      - main
+
+env:
+  ruby_version: 3.3
+
+jobs:
+  rubocop:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Setup Ruby
+        uses: ruby/setup-ruby@v1
+        with:
+          ruby-version: ${{ env.ruby_version }}
+          bundler-cache: true
+
+      - uses: actions/checkout@v4
+
+      - name: Install gems
+        run: bundle install
+
+      - name: Rubocop
+        run: bundle exec rubocop
