+++ .rubocop.yml
@@ -10,6 +10,7 @@
 # See https://docs.rubocop.org/rubocop/configuration
 AllCops:
   NewCops: enable
+  TargetRubyVersion: 2.5
 
 Metrics/BlockLength:
   Exclude:
